Extract inline ArrayAdapter to private MemberMenuAdapter in MembersAdapter
Replaced the inline `ArrayAdapter` used for the AlertDialog overflow menu
in `MembersAdapter` with a private class named `MemberMenuAdapter`.
This improves code readability, cleanly encapsulates the list styling
logic within the same file, and avoids package namespace pollution.
